+++ R/osmdata-package.R
@@ -16,7 +16,7 @@
 #' (or an unnamed vector) return a string
 #' \item \code{\link{overpass_status}}: Retrieve status of the overpass API
 #' \item \code{\link{opq}}: Build an overpass query
-#' \item \code{\link{add_feature}}: Add a feature to an overpass query
+#' \item \code{\link{add_osm_feature}}: Add a feature to an overpass query
 #' \item \code{\link{opq_string}}: Convert an osmdata query to overpass API
 #' string
 #' }